voting cannot be spammed
