+++ src/components/SingleReview.jsx
@@ -5,7 +5,6 @@ import { formatDate } from "../utils/utils";
 import { UserIcon } from "./UserIcon";
 import { Loading } from "./Loading";
 import { Comments } from "./Comments";
-import { patchReview } from "../utils/api-requests";
 import { Error } from "./Error";
 import {ReviewVoting} from "./ReviewVoting"
 import {ReviewComment} from './ReviewComment'
@@ -14,7 +13,7 @@ export const SingleReview = () => {
   const [review, setReview] = useState({});
   const [isLoading, setIsLoading] = useState(true);
   const [isHidden, setIsHidden] = useState(true);
-  const [localVotes, setLocalVotes] = useState("");
+
   const [err, setErr] = useState(null);
 
   const { review_id } = useParams();
@@ -34,24 +33,8 @@ export const SingleReview = () => {
     getReview(review_id).then((reviewFromApi) => {
       setReview(reviewFromApi);
       setIsLoading(false);
-      setLocalVotes(votes);
     });
   }, [review_id, votes, isHidden]);
-
-  const handleVoteClick = (event) => {
-    event.preventDefault();
-    setErr(null);
-    setLocalVotes(localVotes + Number(event.target.value));
-    patchReview(review_id, event.target.value)
-      .catch((err) => {
-        setLocalVotes(localVotes - Number(event.target.value));
-        setErr("Something went wrong, please try again.");
-      });
-  };
-
-  const handleCommentClick = () => {
-    setIsHidden(false);
-  };
 
   return isLoading ? (
     <Loading />
@@ -70,8 +53,8 @@ export const SingleReview = () => {
         <p>{formatDate(created_at)}</p>
         <p>{review_body}</p>
         <span>
-          <ReviewVoting handleVoteClick={handleVoteClick} localVotes={localVotes}/> 
-          <ReviewComment handleCommentClick={handleCommentClick} comment_count = {comment_count}/>
+          <ReviewVoting review_id= {review_id} setErr= {setErr} votes={votes}/> 
+          <ReviewComment setIsHidden = {setIsHidden} comment_count = {comment_count}/>
 
         </span>
         {isHidden ? null : <Comments setIsHidden={setIsHidden} />}
